chore: move pod and pvc events function to k8stest package
