+++ common/mayastor/restore/restore.go
@@ -2,12 +2,9 @@ package restore
 
 import (
 	"fmt"
-	"strings"
-	"time"
 
 	"github.com/openebs/openebs-e2e/common"
 	"github.com/openebs/openebs-e2e/common/controlplane"
-	"github.com/openebs/openebs-e2e/common/k8stest"
 
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 )
@@ -29,34 +26,6 @@ var (
 	VolumeModeConversionAnnotationValue    = "true"
 	ClonedSnapshotVolumeThinErrorsubstring = "ClonedSnapshotVolumeThin: Cloned snapshot volumes must be thin provisioned"
 )
-
-func IsWarningPvcEventPresent(pvcName string, namespace string, errorSubstring string) (bool, error) {
-	events, err := k8stest.GetPvcEvents(pvcName, namespace)
-	if err != nil {
-		return false, fmt.Errorf("failed to get pvc events in namespace %s, error: %v", namespace, err)
-	}
-	for _, event := range events.Items {
-		logf.Log.Info("Found PVC event", "message", event.Message)
-		if event.Type == "Warning" && strings.Contains(event.Message, errorSubstring) {
-			return true, err
-		}
-	}
-	return false, err
-}
-
-func IsWarningPodEventPresent(podName string, namespace string, errorSubstring string) (bool, error) {
-	events, err := k8stest.GetPodEvents(podName, namespace)
-	if err != nil {
-		return false, fmt.Errorf("failed to get pod events in namespace %s, error: %v", namespace, err)
-	}
-	for _, event := range events.Items {
-		logf.Log.Info("Found Pod event", "message", event.Message)
-		if event.Type == "Warning" && strings.Contains(event.Message, errorSubstring) {
-			return true, err
-		}
-	}
-	return false, err
-}
 
 func VerifyRestoredVolumeList(restVol map[string]common.Snapshot) (bool, error) {
 	logf.Log.Info("Verify restored volumes list via control plane")
@@ -89,22 +58,4 @@ func VerifyRestoredVolumeList(restVol map[string]common.Snapshot) (bool, error) 
 	}
 	return true, err
 
-}
-
-// Wait for the PVC warning event
-// return true if pvc warning event found else return false , it return error in case of error
-func WaitForPvcWarningEvent(pvcName string, namespace string, errorSubstring string) (bool, error) {
-	const timeSleepSecs = 1
-	var hasWarning bool
-	var err error
-	// Wait for the PVC event
-	logf.Log.Info("Check Pvc event", "Event substring", errorSubstring)
-	for ix := 0; ix < DefTimeoutSecs/timeSleepSecs; ix++ {
-		hasWarning, err = IsWarningPvcEventPresent(pvcName, namespace, errorSubstring)
-		if err == nil && hasWarning {
-			break
-		}
-		time.Sleep(timeSleepSecs * time.Second)
-	}
-	return hasWarning, err
 }
